requirements.txt
streamlit>=1.30.0
pandas>=2.0.0
numpy>=1.24.0
plotly>=5.18.0
openpyxl>=3.1.0
reportlab>=4.0.0
bcrypt>=4.1.0
cryptography>=42.0.0
python-dotenv>=1.0.0
